Add reset weights endpoint (#486)
* Implement reset weights endpoint

* Cleanup client routes
